Replace quick failure mechanism with exponential backoff
Fundamentally, we do not want or anticipate our watchdog to ever stop,
for any reason. The quick failure mechanism we have implemented moves
counter to this desire, albeit more or less intentionally so. The idea
behind it was that if a program is misconfigured it may fail very
quickly very early on and we can easily detect that. Unfortunately,
there may also be legitimate cases where we could see such a pattern:
consider a program that tries to connect to a server as its first
action, and the ISP has temporarily cut the Internet connection.
Ultimately, the believe is that there is no "safe" quick failure count.
Hence, with this change we go down a different road: we instead just
use an exponential back off when restarting the managed process.