Use mio crate and SIGCHLD events for awaiting child exit
The way we currently wait for the child to exit does not play nice with
any attempts of gracefully terminating the watchdog by the user, e.g.,
via Ctrl-C. At the same time, we have plans for reacting on similar
additional events in the future.
To future proof the code for both possibilities, switch over to rolling
a proper event loop using the mio crate. Instead of directly blocking
waiting for the child to exit, register a SIGCHLD handler and register a
poll waiting for it to trigger.