Attempt to gracefully exit program on Ctrl-C
So far our handling of Ctrl-C, which maps to SIGINT, has been the
default: the program gets terminated. The problem with this approach is
that the watched child can leak in the process, which is not desirable
behavior.
With this change we register a SIGINT handler and poll for events on it.
Once received for the first time, we attempt to gracefully shutdown the
child by sending it a SIGTERM. If a second SIGINT is sent, we forcibly
kill the child and then exit immediately.