Gracefully handle SIGTERM signals
Gracefully handle SIGTERM signals, in a way similar but not quite the
same as SIGINT. For SIGTERM we do the same graceful termination
procedure as for SIGINT, but instead of waiting for a second signal to
force the kill we automate this step after a timeout of 500ms. A
procedure like this arguably makes the most sense for SIGTERM, which
typically is sent by automation or with the expectation to more
definitely terminate the process in the near future.