Introduce a proper state machine
Our event loop is written in somewhat of an ad-hoc fashion, without
making some possible states explicit. This can make it hard to reason
about state transitions and it also leads to weird behavior such as
forced termination not working while a backoff is in progress.
With this change we convert the logic into an explicit state machine,
formalizing the three different states identified (running, signaled for
termination, and backing off) and establishing proper transitions
between them.