Forget signal handler pipe endpoints on program exit
We have seen cases where our signal handlers fail to write data to the
pipe connecting them to the rest of the program on program shutdown. The
reason seems to be that the main thread already closed the receiving
ends of said pipes while a signal got triggered.
There is no real reason for us to close these pipe ends on what is
already a path where the operating system claims back control of all
resources. Hence, in an attempt to prevent these issues from potentially
showing up in the future, just "forget" the corresponding pipe ends.