Add logic for "rotating" files
Watched processes are often used in service-style settings, where output
data is ultimately forwarded into a log file of sorts. For long running
processes this can ultimately end up being a problem, if data is
retained unconditionally forever, as file systems will inevitably fill
out.
To counter such behavior, a common strategy is to "rotate" log files
after they reach a certain size, keeping a certain number of "archived"
files around, but discarding them when newer data is available.
This change introduces the infrastructure for rotating logs to the
program. At the moment this is not yet hooked up to the core logic, but
that will change in the future.