Wire up file rotation facility
Wire up the file rotation facility with the rest of the program, by
adding necessary new program arguments to control which streams to
capture and which log file to write them out to, as well as limits on
maximum number of files as well as lines in them. Everything is
integrated with the existing event loop and, hence, is fully event
driven.