Provide custom logger implementation
So far we have used the env_logger crate for initializing the logging
infrastructure that helps provide some level of introspection for what
the watchdog is doing. However, the reliance on the RUST_LOG environment
variable is problematic: it may very well be used by the watched process
as well, in which case it is impossible (well, unnecessarily hard) to
control verbosity of the two processes in isolation.
With this change we control the program's verbosity solely via the newly
introduced command line argument -v/--verbose. In the process we switch
away from env_logger entirely and provide our own (trivial) log::Log
implementation, over which we have full control. In the future, we may
opt for forwarding the watchdog's log messages into the rotated log file
along with the watched process' output.