Move Rotate object management into Watchdog type
When we married the file rotation facility with the rest of the logic we
co-located them inside the Watched type, as that was where streams for
the spawned process needed to be configured. However, in retrospect that
location was sub-optimal: not only does it mean that we have to plumb
through all the arguments necessary for setting up the rotation
facility, we also recreate it whenever the watched process is restarted.
It also means that there is no program-wide object representing the
rotated log, which effectively makes it impossible to forward the
watchdog's own logs to it.
With this change we move the logic for managing the Rotate instance into
the Watchdog type itself, which greatly simplifies the Watched type
setup and will allow us to access the log from the core watchdog logic.