Support routing log messages into rotated log file
So far our program's log messages were strictly separate from that of
the watched process. That may or may not be desired. In a service style
setting, it would likely be best to have the watchdog's messages go into
the same log file (if configured) that the watched process' output goes
into.
With this change we add the necessary infrastructure in support of this
goal. It is somewhat unfortunately that the log crate requires the
set logger to be Send + Sync, and so we have to enforce synchronization
even in a single threaded context, but it is probably not worth it to
reimplement the logging infrastructure at a lower level as well.